test update

+++ tests/samples/Remove-MyApp.ps1
@@ -1,5 +1,8 @@
 $authServer = Get-OktaAuthorizationServer -Query 'Casualty-Myapp-Test-AS'
 $application = Get-OktaApplication -Query 'Casualty-Myapp-Test'
-
-Remove-OktaAuthorizationServer -AuthorizationServerId $authServer.id -Confirm:$false
-Remove-OktaApplication -AppId $application.id -Confirm:$false
+if ($authServer) {
+    Remove-OktaAuthorizationServer -AuthorizationServerId $authServer.id -Confirm:$false
+}
+if ($application) {
+    Remove-OktaApplication -AppId $application.id -Confirm:$false
+}
